+++ include/system_solver.h
@@ -5,25 +5,25 @@
 class SystemSolver {
 public:
   // Return empty vector if method have error.
-  void Solve(const std::vector<float>& upper_diag,
-             const std::vector<float>& main_diag,
-             const std::vector<float>& lower_diag,
-             const std::vector<float>& right_part,
-             std::vector<float>* y);
+  void Solve(const std::vector<double>& upper_diag,
+             const std::vector<double>& main_diag,
+             const std::vector<double>& lower_diag,
+             const std::vector<double>& right_part,
+             std::vector<double>* y);
 private:
   void ForwardMove();
 
-  void BackwardMove(std::vector<float>* y);
+  void BackwardMove(std::vector<double>* y);
 
   int n_;
-  float kappa_1_;
-  float kappa_2_;
-  float mu_1_;
-  float mu_2_;
-  std::vector<float> phi_;
-  std::vector<float> A_;
-  std::vector<float> B_;
-  std::vector<float> C_;
-  std::vector<float> alpha_;
-  std::vector<float> beta_;
+  double kappa_1_;
+  double kappa_2_;
+  double mu_1_;
+  double mu_2_;
+  std::vector<double> phi_;
+  std::vector<double> A_;
+  std::vector<double> B_;
+  std::vector<double> C_;
+  std::vector<double> alpha_;
+  std::vector<double> beta_;
 };
